work on ecnf upload scripts + new TraceHash functions in lmfdb/utils
