+++ lmfdb/utils/__init__.py
@@ -42,7 +42,8 @@ __all__ = ['request', 'make_response', 'flash', 'url_for', 'render_template',
            'Configuration',
            'names_and_urls', 'name_and_object_from_url',
            'datetime_to_timestamp_in_ms', 'timestamp_in_ms_to_datetime',
-           'reraise']
+           'reraise',
+           'TraceHash', 'TraceHashClass']
 
 from flask import (request, make_response, flash, url_for,
                    render_template, send_file)
@@ -89,3 +90,4 @@ from .display_stats import formatters, proportioners, totaler, StatsDisplay
 from .config import Configuration
 from .names_and_urls import names_and_urls, name_and_object_from_url
 from .reraise import reraise
+from .trace_hash import TraceHash, TraceHashClass
